Windows (MinGW-w64) support (#2)
Build and run propwash on Windows via MinGW-w64 (MSYS2), verified on a
windows-latest CI runner. Betaflight SITL, the sim, the UDP wire protocol
and the CLI/MSP all work; the whole firmware compiles under MinGW.

Build/link:
- BF_LINK_OPTIONS WIN32 branch: pg.ld + --gc-sections (MinGW ld), -lpthread,
  no -lrt; ws2_32 declared as a PUBLIC dep of bf_sitl so CMake orders Winsock
  after the referencing objects.
- pw_stubs.c (mpuGyroReadRegister) reused on MinGW — --gc-sections doesn't
  prune the dead gyroReadRegister there either.
- Compat shims, Windows include path only: an empty <netinet/in.h> (BF's
  serial_tcp.h includes it vestigially; forwarding to Winsock would drag in
  windows.h, whose BAUD_* macros collide with baudRate_e), a dyad.h wrapper
  that #include_next's the stock header then undefs those BAUD_* macros, and
  pw_win_compat.h for ffs/strcasestr/<string.h>.

Runtime fixes (real bugs, not just build plumbing):
- core/net/server.cpp: Winsock port (WSAStartup, SO_RCVTIMEO as DWORD ms,
  char* buffers, SOCKET in an intptr_t).
- serial_tcp.c: bind the CLI listener on IPv4 explicitly. Windows
  getaddrinfo(NULL,...) returns IPv6 first and v6 sockets are v6-only, so the
  CLI bound [::]:5761 and refused the 127.0.0.1 everything actually uses.
- pw_sitl.c: open the eeprom file in BINARY mode. Text mode was translating
  every 0x0A in the config image to CRLF, corrupting it so PG records failed
  their checksum and loaded as defaults.
- pw_cli.py: flow-controlled _drain (wait for the "# " prompt), a single-poke
  enter() that flushes trailing prompts, and connection retry. These make the
  CLI robust on slow hosts and are improvements on every platform.

CI: build-and-test-windows job (MSYS2 MINGW64, Ninja, native MinGW python3).

KNOWN LIMITATION: diff_roundtrip and real_tune_hover are skipped on Windows.
Both reconnect to the CLI after the firmware reboots, and that reconnect
intermittently hangs there (the dyad listener stops accepting). Suspected
cause — stale tcpPort_t client state surviving an in-process systemReset — is
documented inline in ci.yml. Both still run on Linux and macOS.

+++ core/net/server.h
@@ -31,7 +31,9 @@ class Server {
   private:
     void sendTo(const void* payload, uint16_t len, uint8_t type);
 
-    int mFd = -1;
+    // Holds a POSIX fd (int) or a Windows SOCKET (UINT_PTR). intptr_t fits
+    // both, and INVALID_SOCKET (~0) lands as -1, so `< 0` stays the error test.
+    intptr_t mFd = -1;
     // last known client
     uint32_t mClientAddr = 0;
     uint16_t mClientPort = 0;
